+++ src/containers/Social/SocialLinks.tsx
@@ -1,4 +1,4 @@
-import { socialSection } from '@/lib/content/socialSection';
+import { socialSection } from '@/lib/content/social';
 
 import { SocialIcon } from '@/components';
