+++ src/lib/content/footer.ts
