+++ src/lib/content/hero.ts
