+++ src/lib/content/navbar.ts
